Added the concept of "updaters" + a thermostat
An Updater is something that can modify the system, either acting on the
particles or on the system properties (such as the temperature or the
box). System stores a list of updaters that are called at the end of
each time step.
Thermostats are an example of updaters. This commit introduces the first
thermostat in the form of a modified Andersen thermostat.