Removed a memory leak due to shared_ptr's and cross-dependencies
ForceComputer's are stored in shared_ptr's that were originally kept in
the SystemProperties class. However, ForceComputer's (and Observable's)
also keep shared_ptr's to the SystemProperties class. This
cross-dependency was generating memory leaks, since shared_ptr's would
not be correctly clean up. The ForceComputer's pointers have been moved
to the System class to circumvent the issue.